More appeasements to phpstan, taking the number of reported errors down below 175

+++ src/PhpSpreadsheet/Calculation/Financial/Coupons.php
@@ -293,7 +293,7 @@ class Coupons
 
         $yearsBetweenSettlementAndMaturity = DateTimeExcel\YearFrac::funcYearFrac($settlement, $maturity, 0);
 
-        return ceil($yearsBetweenSettlementAndMaturity * $frequency);
+        return (int) ceil($yearsBetweenSettlementAndMaturity * $frequency);
     }
 
     /**
